requirements.txt
python_vlc~=3.0.12118
fpstimer~=0.0.1
opencv_python~=4.5.2.52
Pillow~=8.2.0
colorama~=0.4.4
